Moved some work into new ProcessExtensions; reduced SetGameProcessAndVersion.
